add a recording system, restructure headers
This adds the recording system, all pressed keys are recorded besides
normal mode motions. This is used to implement the `.` keybind to repeat
the last action and soon for repeated insertion but that is not included
in this commit.

The header files are restructured, `mode.c` and `mode.h` are gone,
`mode.h` is now `purec.h` and the `mode.c` code got moved into `main()`
and trimmed down.

The global variable `Mode` is now `Core`.