add multi insert, include `break_line()`
This adds multi line insertions (again) for visual block mode when
pressing `I` or `A`. It also re-adds repeated insertion, so doing `4i`
in normal mode to insert the following text 4 times (C-C to cancel).

It also adds `break_line()` that will be used to indent properly but the
c language parser needs to have bracket matching first.

Transient event chains now are like they used to be with the start
having the valid `undo_cur` and the end the valid `redo_cur`.