add frame motions
The "frame motions" are aware of the height of the frame, use `H` to go
to the beginning of the frame, `M` to go the the middle and `L` to reach to the
end. `H` and `L` also move up and down respectively if they are already
at the begginning/end of the frame.