change undo system
Now the undo system works with a pool of text segments and each event
gets an index into that. This is so duplicate text segments do not get
stored twice. This also allows to use a file off memory to not put a too
heavy load on memory when there are very large undoes.

This system needs refining..