rework `do_motion()`, improve indentation further
Now the frame stores the information of a motion, where the future
vct is and the cursor when calling `prepare_motion()`. If this
function returns 1, then `apply_motion()` can be called to do the
motion.

This system fixes a lot of issues where the frame scrolls away,
however, a lot of testing is missing.